# docs/swagger.yaml
definitions:
  request.DeployUpdateMessage:
    properties:
      name:
        type: string
      namespace:
        type: string
      newImage:
        type: string
      replicasNumber:
        type: integer
    type: object
  request.DeploymentRequest:
    properties:
      namespace:
        type: string
      object:
        additionalProperties: true
        type: object
    type: object
...
  response.CommonResponse:
    properties:
      data: {}
      msg:
        type: string
    type: object
...
info:
  contact: {}
  description: This is a k8s cluster management service
  title: K8S Client API
  version: 0.0.1
paths:
...
  /deployment/create:
    post:
      parameters:
      - description: Deployment simple configuration
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.DeploymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.CommonResponse'
      security:
      - ApiKeyAuth: []
      summary: 创建 Deployment
      tags:
      - Deployment
  /deployment/delete:
    delete:
      parameters:
      - default: default
        description: 命名空间
        in: query
        name: namespace
        type: string
      - description: deployment 名称
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.CommonResponse'
      security:
      - ApiKeyAuth: []
      summary: 删除单个 Deployment
      tags:
      - Deployment
  /deployment/get:
    get:
      parameters:
      - default: default
        description: 命名空间
        in: query
        name: namespace
        type: string
      - description: deployment 名称
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.CommonResponse'
      security:
      - ApiKeyAuth: []
      summary: 获取单个 Deployment
      tags:
      - Deployment
  /deployment/list:
    get:
      parameters:
      - default: default
        description: 命名空间
        in: query
        name: namespace
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.CommonResponse'
      security:
      - ApiKeyAuth: []
      summary: 获取命名空间下的所有 Deployment
      tags:
      - Deployment
  /deployment/update:
    put:
      parameters:
      - description: Deployment configuration information that needs to be changed
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.DeployUpdateMessage'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.CommonResponse'
      security:
      - ApiKeyAuth: []
      summary: 更新 Deployment 的镜像版本和副本集
      tags:
      - Deployment
  /namespace/create:
    post:
      parameters:
      - description: Namespace name
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.CommonResponse'
      security:
      - ApiKeyAuth: []
      summary: 创建 Namespace
      tags:
      - Namespace
  /namespace/delete:
    delete:
      parameters:
      - description: namespace 名称
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.CommonResponse'
      security:
      - ApiKeyAuth: []
      summary: 删除单个 Namespace
      tags:
      - Namespace
  /namespace/get:
    get:
      parameters:
      - description: namespace 名称
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.CommonResponse'
      security:
      - ApiKeyAuth: []
      summary: 获取单个 Namespace
      tags:
      - Namespace
  /namespace/list:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.CommonResponse'
      security:
      - ApiKeyAuth: []
      summary: 获取所有的 Namespace
      tags:
      - Namespace
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"